+++ include/query_exec.h
@@ -9,38 +9,38 @@
 enum operators { EQ, GREATER, LESS };
 
 struct filter {
-  int rel;
-  int col;
+  int32_t rel;
+  int32_t col;
   operators op;
-  int64_t literal;
+  int32_t literal;
 
   filter() = default;
 
-  filter(int lr, int lc, operators o, int64_t lit)
+  filter(int32_t lr, int32_t lc, operators o, int32_t lit)
       : rel{lr}, col{lc}, op{o}, literal{lit} {}
 };
 
 struct join {
-  int left_rel;
-  int left_col;
+  int32_t left_rel;
+  int32_t left_col;
   operators op;
-  int right_rel;
-  int right_col;
+  int32_t right_rel;
+  int32_t right_col;
 
   join() = default;
 
-  join(int lr, int lc, operators o, int rr, int rc)
+  join(int32_t lr, int32_t lc, operators o, int32_t rr, int32_t rc)
       : left_rel{lr}, left_col{lc}, op{o}, right_rel{rr}, right_col{rc} {}
 };
 
 struct project_rel {
-  int rel;
-  int col;
+  int32_t rel;
+  int32_t col;
 };
 
 class QueryExec {
-  int qindex;  // which query
-  simple_vector<int> rel_names;
+  int32_t qindex;  // which query
+  simple_vector<int32_t> rel_names;
   simple_vector<join> joins;
   simple_vector<filter> filters;
   simple_vector<project_rel> projections;
@@ -48,8 +48,9 @@ class QueryExec {
   bool rel_is_filtered[4];
   bool rel_is_joined[4];
 
-  simple_vector<int> joined[4];  // Represents latest intermediate results
-  simple_vector<int> filtered[4];
+  simple_vector<int32_t>
+      joined[4];  // Represents latest int32_termediate results
+  simple_vector<int32_t> filtered[4];
 
   statistics *rel_stats[4];
 
@@ -62,7 +63,7 @@ class QueryExec {
   void join_enumeration();
   bool isConnected(simple_vector<join> &, size_t);
   uint64_t calculate_cost(size_t);
-  void update_stats(size_t, int);
+  void update_stats(size_t, int32_t);
 
   void do_query();
   void filter_exec(size_t);
@@ -71,7 +72,7 @@ class QueryExec {
   void checksum();
 
  public:
-  QueryExec(int);
+  QueryExec(int32_t);
   void execute(char *);
 };
